+++ KMeans/KMeans/DeviceRawDataGPU1.cuh
@@ -10,6 +10,6 @@ namespace GPU1
 	template<size_t dim>
 	struct DeviceRawDataGPU1 : public DeviceRawDataGPU<dim> {
 		float* DeviceUpdatedCentroids{};
-		size_t* DeviceUpdatedCentroidsCounts{};
+		unsigned* DeviceUpdatedCentroidsCounts{};
 	};
 }
